requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
google-generativeai==0.7.2 
gunicorn==21.2.0
requests==2.31.0
